worked on comments from @sandjelkovic
